Consolidate EdgeCases into StaticCallOutsideClass
