--- tests/Fixtures/src/TypeInference/EdgeCases.php
@@ -1,20 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Fixtures\TypeInference;
-
-function thisInFunction(): void
-{
-    $x = $this;
-}
-
-function unknownClassParameter(UnknownClass $obj): void
-{
-    $result = $obj->name;
-}
-
-function methodCallOnUnresolvedType(): void
-{
-    $result = $unknown->someMethod();
-}
