+++ tests/Fixtures/src/TypeInference/StaticCallOutsideClass.php
@@ -13,3 +13,18 @@ function callParentOutsideClass(): void
 {
     $x = parent::method();
 }
+
+function thisInFunction(): void
+{
+    $x = $this;
+}
+
+function unknownClassParameter(UnknownClass $obj): void
+{
+    $result = $obj->name;
+}
+
+function methodCallOnUnresolvedType(): void
+{
+    $result = $unknown->someMethod();
+}
